Driver now checks for hidden.
the driver will not check whether or not hide was called or not, before running show, if hide was not called, show will not run. this is to make sure self.show is not called on the remote when it is already showing as that could cause weird behaviour maybe.